save progress

+++ server/server.go
@@ -5,15 +5,15 @@ import (
 	"database/sql"
 	"fmt"
 	"net/url"
-	"time"
 	"strings"
+	"time"
 
 	"bitbucket.org/exonch/resource-service/server/model"
 	"bitbucket.org/exonch/resource-service/server/other"
 	"bitbucket.org/exonch/resource-service/util/cache"
 
-	"github.com/sirupsen/logrus"
 	uuid "github.com/satori/go.uuid"
+	"github.com/sirupsen/logrus"
 )
 
 var _ = logrus.StandardLogger()
@@ -22,6 +22,13 @@ type ResourceSvcInterface interface {
 	CreateNamespace(ctx context.Context, userID, nsLabel, tariffID string, adminAction bool) error
 	DeleteNamespace(ctx context.Context, userID, nsLabel string) error
 	ListNamespaces(ctx context.Context, userID string, adminAction bool) ([]Namespace, error)
+	GetNamespace(ctx context.Context, userID, nsLabel string, adminAction bool) (Namespace, error)
+	ChangeAccessToNamespace(ownerUserID, nsLabel string, otherUserID, accessLevel string) error
+	LockAccessToNamespace(ownerUserID, nsLabel string, lockState bool) error
+
+	//CreateExtService(ctx context.Context, userID, svLabel string, adminAction bool) (Service, error)
+	//DeleteExtService(ctx context.Context, userID, svLabel string, adminAction bool) error
+	//ListExtServices(ctx context.Context, userID string, adminAction bool) error
 }
 
 type ResourceSvc struct {
@@ -84,6 +91,8 @@ func (rm *ResourceSvc) CreateNamespace(ctx context.Context, userID, nsLabel, tar
 		}
 	}
 
+	err = rm.db.namespaceCreate(resourceUUDI
+
 	err = rm.db.permCreate("Namespace", resourceUUID, userUUID)
 	if err != nil {
 		return newError("database error")
@@ -92,7 +101,8 @@ func (rm *ResourceSvc) CreateNamespace(ctx context.Context, userID, nsLabel, tar
 	ctx, cancelf := context.WithCancel(ctx)
 	waitch := make(chan struct{})
 	go func() {
-		errKube = rm.kube.CreateNamespace(ctx, resourceUUID.String(), uint(*tariff.CpuLimit), uint(*tariff.MemoryLimit))
+		errKube = rm.kube.CreateNamespace(ctx, resourceUUID.String(), uint(*tariff.CpuLimit), uint(*tariff.MemoryLimit),
+			nsLabel, "owner")
 		if errKube != nil {
 			cancelf()
 		}
@@ -108,30 +118,32 @@ func (rm *ResourceSvc) CreateNamespace(ctx context.Context, userID, nsLabel, tar
 	<-waitch
 	<-waitch
 
-	go func() {
-		if errKube == nil {
-			rm.kube.DeleteNamespace(context.Background(), resourceUUID.String())
+	if errKube != nil || errBilling != nil {
+		go func() {
+			if errKube == nil {
+				rm.kube.DeleteNamespace(context.Background(), resourceUUID.String())
+			}
+			if errBilling == nil {
+				logrus.Warnf("would unsubscribe user %q from tariff %q", userID, tariffID)
+				rm.billing.Unsubscribe(context.Background(), userID, resourceUUID.String())
+			}
+		}()
+		var errs []string
+		if errKube != nil {
+			errs = append(errs, fmt.Sprintf("kube api error: %v", errKube))
 		}
-		if errBilling == nil {
-			logrus.Warnf("would unsubscribe user %q from tariff %q", userID, tariffID)
-			rm.billing.Unsubscribe(context.Background(), userID, resourceUUID.String())
+		if errBilling != nil {
+			errs = append(errs, fmt.Sprintf("billing error: %v", errBilling))
 		}
-	}()
-
-	var errs []string
-	if errKube != nil {
-		errs = append(errs, fmt.Sprintf("kube api error: %v", errKube))
-	}
-	if errBilling != nil {
-		errs = append(errs, fmt.Sprintf("billing error: %v", errBilling))
-	}
-	if len(errs) > 0 {
-		err = newOtherServiceError("%s", strings.Join(errs, "; "))
-		return err
+		if len(errs) > 0 {
+			err = newOtherServiceError("%s", strings.Join(errs, "; "))
+			return err
+		}
+	} else {
+		go rm.mailer.SendNamespaceCreated(model.User{ID: &userID}, nsLabel, model.Tariff{ID: &tariffID})
+		go rm.authsvc.UpdateUserAccess(userID)
 	}
 
-	go rm.mailer.SendNamespaceCreated(model.User{ID: &userID}, nsLabel, model.Tariff{ID: &tariffID})
-	go rm.authsvc.UpdateUserAccess(userID)
 	return nil
 }
 
@@ -154,6 +166,93 @@ func (rm *ResourceSvc) ListNamespaces(ctx context.Context, userID string, adminA
 		}
 	}
 	return namespaces, nil
+}
+
+func (rs *ResourceSvc) GetNamespace(ctx context.Context, userID, nsLabel string, adminAction bool) (ns Namespace, err error) {
+	userUUID, err := uuid.FromString(userID)
+	if err != nil {
+		err = newBadInputError("invalid user id %q: %v", userID, err)
+		return
+	}
+	ns, err = rs.db.namespaceGet(userUUID, nsLabel)
+	if err != nil {
+		err = newError("database: %v", err)
+		return
+	}
+	if !adminAction {
+		ns.ID = nil
+	}
+	return
+}
+
+// ChangeAccessToNamespace adds or removes access to ownerUserID's resource for
+// otherUserID.
+func (rs *ResourceSvc) ChangeAccessToNamespace(ownerUserID, nsLabel string, otherUserID, permOther string) error {
+	ownerUserUUID, err := uuid.FromString(ownerUserID)
+	if err != nil {
+		return newBadInputError("invalid ownerUserID: %v", err)
+	}
+	otherUserUUID, err := uuid.FromString(otherUserID)
+	if err != nil {
+		return newBadInputError("invalid otherUserID: %v", err)
+	}
+	ns, err := rs.db.namespaceGet(ownerUserUUID, nsLabel)
+	if err != nil {
+		switch err.(type) {
+		case Error:
+			return err
+		default:
+			return newError("database: %v", err)
+		}
+	}
+
+	perm, err := rs.db.permFetch(*ns.ID, ownerUserUUID)
+	if err != nil {
+		return newError("database: %v", err)
+	}
+	if permCheck(perm, "owner") == false {
+		return newPermissionError("permission denied")
+	}
+
+	if err = rs.db.permSetOtherUser(*ns.ID, otherUserUUID, permOther); err != nil {
+		switch err.(type) {
+		case Error, BadInputError:
+			return err
+		default:
+			return newError("database: %v", err)
+		}
+	}
+
+	return nil
+}
+
+func (rs *ResourceSvc) LockAccessToNamespace(ownerUserID, nsLabel string, lockState bool) error {
+	ownerUserUUID, err := uuid.FromString(ownerUserID)
+	if err != nil {
+		return newBadInputError("invalid owner user id: %v", err)
+	}
+	ns, err := rs.db.namespaceGet(ownerUserUUID, nsLabel)
+	if err != nil {
+		switch err.(type) {
+		case Error:
+			return err
+		default:
+			return newError("database: %v", err)
+		}
+	}
+	perm, err := rs.db.permFetch(*ns.ID, ownerUserUUID)
+	if err != nil {
+		return newError("database: %v", err)
+	}
+	if permCheck(perm, "owner") == false {
+		return newPermissionError("permission denied")
+	}
+
+	if err = rs.db.permSetLimited(*ns.ID, lockState); err != nil {
+		return newError("database: %v", err)
+	}
+
+	return nil
 }
 
 func (rm *ResourceSvc) getNSTariff(ctx context.Context, id string) (t model.NamespaceTariff, err error) {
